+++ R/dbplyr.R
@@ -12,14 +12,6 @@
 "_PACKAGE"
 
 # Generics that really should live in dbplyr
-#' @importFrom dplyr db_analyze
-#' @importFrom dplyr db_create_index
-#' @importFrom dplyr db_desc
-#' @importFrom dplyr db_explain
-#' @importFrom dplyr db_query_fields
-#' @importFrom dplyr db_query_rows
-#' @importFrom dplyr db_save_query
-#' @importFrom dplyr db_write_table
 #' @importFrom dplyr sql_join
 #' @importFrom dplyr sql_select
 #' @importFrom dplyr sql_semi_join