+++ R/tbl-sql.R
@@ -16,7 +16,7 @@ tbl_sql <- function(subclass, src, from, ..., vars = NULL) {
   # If not literal sql, must be a table identifier
   from <- as.sql(from, con = src$con)
 
-  vars <- vars %||% db_query_fields(src$con, from)
+  vars <- vars %||% dplyr::db_query_fields(src$con, from)
   ops <- op_base_remote(from, vars)
 
   dplyr::make_tbl(c(subclass, "sql", "lazy"), src = src, ops = ops)
@@ -72,7 +72,7 @@ tbl_sum.tbl_sql <- function(x) {
   sort <- op_sort(x$ops)
   c(
     "Source" = tbl_desc(x),
-    "Database" = db_desc(x$src$con),
+    "Database" = dplyr::db_desc(x$src$con),
     "Groups" = if (length(grps) > 0) commas(grps),
     "Ordered by" = if (length(sort) > 0) commas(deparse_all(sort))
   )